feat(server): add transport mode
