+++ lib/Server.js
@@ -61,15 +61,29 @@ class Server {
 
     this.log = _log || createLogger(options);
 
-    if (this.options.serverMode !== undefined) {
-      this.log.warn(
-        'serverMode is an experimental option, meaning its usage could potentially change without warning'
-      );
-    }
+    if (this.options.transportMode === undefined) {
+      this.options.transportMode = {
+        server: 'sockjs',
+        client: 'sockjs',
+      };
+    } else {
+      switch (typeof this.options.transportMode) {
+        case 'string':
+          this.options.transportMode = {
+            server: this.options.transportMode,
+            client: this.options.transportMode,
+          };
+          break;
+        // if not a string, it is an object
+        default:
+          this.options.transportMode.server =
+            this.options.transportMode.server || 'sockjs';
+          this.options.transportMode.client =
+            this.options.transportMode.client || 'sockjs';
+      }
 
-    if (this.options.clientMode !== undefined) {
       this.log.warn(
-        'clientMode is an experimental option, meaning its usage could potentially change without warning'
+        'transportMode is an experimental option, meaning its usage could potentially change without warning'
       );
     }
 
@@ -674,7 +688,7 @@ class Server {
 
       if (!headers) {
         this.log.warn(
-          'serverMode implementation must pass headers to the callback of onConnection(f) ' +
+          'transportMode.server implementation must pass headers to the callback of onConnection(f) ' +
             'via f(connection, headers) in order for clients to pass a headers security check'
         );
       }
