+++ lib/utils/getSocketClientPath.js
@@ -3,17 +3,17 @@
 function getSocketClientPath(options) {
   let ClientImplementation;
   let clientImplFound = true;
-  switch (typeof options.clientMode) {
+  switch (typeof options.transportMode.client) {
     case 'string':
       // could be 'sockjs', 'ws', or a path that should be required
-      if (options.clientMode === 'sockjs') {
+      if (options.transportMode.client === 'sockjs') {
         ClientImplementation = require('../../client/clients/SockJSClient');
-      } else if (options.clientMode === 'ws') {
+      } else if (options.transportMode.client === 'ws') {
         ClientImplementation = require('../../client/clients/WebsocketClient');
       } else {
         try {
           // eslint-disable-next-line import/no-dynamic-require
-          ClientImplementation = require(options.clientMode);
+          ClientImplementation = require(options.transportMode.client);
         } catch (e) {
           clientImplFound = false;
         }
@@ -25,7 +25,7 @@ function getSocketClientPath(options) {
 
   if (!clientImplFound) {
     throw new Error(
-      "clientMode must be a string denoting a default implementation (e.g. 'sockjs', 'ws') or a full path to " +
+      "transportMode.client must be a string denoting a default implementation (e.g. 'sockjs', 'ws') or a full path to " +
         'a JS file which exports a class extending BaseClient (webpack-dev-server/client-src/clients/BaseClient) ' +
         'via require.resolve(...)'
     );
