+++ lib/utils/getSocketServerImplementation.js
@@ -3,17 +3,17 @@
 function getSocketServerImplementation(options) {
   let ServerImplementation;
   let serverImplFound = true;
-  switch (typeof options.serverMode) {
+  switch (typeof options.transportMode.server) {
     case 'string':
       // could be 'sockjs', in the future 'ws', or a path that should be required
-      if (options.serverMode === 'sockjs') {
+      if (options.transportMode.server === 'sockjs') {
         ServerImplementation = require('../servers/SockJSServer');
-      } else if (options.serverMode === 'ws') {
+      } else if (options.transportMode.server === 'ws') {
         ServerImplementation = require('../servers/WebsocketServer');
       } else {
         try {
           // eslint-disable-next-line import/no-dynamic-require
-          ServerImplementation = require(options.serverMode);
+          ServerImplementation = require(options.transportMode.server);
         } catch (e) {
           serverImplFound = false;
         }
@@ -22,7 +22,7 @@ function getSocketServerImplementation(options) {
     case 'function':
       // potentially do more checks here to confirm that the user implemented this properlly
       // since errors could be difficult to understand
-      ServerImplementation = options.serverMode;
+      ServerImplementation = options.transportMode.server;
       break;
     default:
       serverImplFound = false;
@@ -30,7 +30,7 @@ function getSocketServerImplementation(options) {
 
   if (!serverImplFound) {
     throw new Error(
-      "serverMode must be a string denoting a default implementation (e.g. 'sockjs', 'ws'), a full path to " +
+      "transportMode.server must be a string denoting a default implementation (e.g. 'sockjs', 'ws'), a full path to " +
         'a JS file which exports a class extending BaseServer (webpack-dev-server/lib/servers/BaseServer) ' +
         'via require.resolve(...), or the class itself which extends BaseServer'
     );
