+++ test/e2e/TransportMode.test.js
@@ -3,33 +3,35 @@
 const testServer = require('../helpers/test-server');
 const config = require('../fixtures/client-config/webpack.config');
 const runBrowser = require('../helpers/run-browser');
-const port = require('../ports-map').ClientMode;
+const port = require('../ports-map').TransportMode;
 const {
   initConsoleDelay,
   awaitServerCloseDelay,
 } = require('../helpers/puppeteer-constants');
 
-describe('clientMode', () => {
+describe('transportMode client', () => {
   const modes = [
     {
       title: 'sockjs',
       options: {
-        clientMode: 'sockjs',
+        transportMode: 'sockjs',
       },
     },
     {
       title: 'ws',
       options: {
-        clientMode: 'ws',
-        serverMode: require.resolve('../../lib/servers/WebsocketServer'),
+        transportMode: 'ws',
       },
     },
     {
       title: 'custom client',
       options: {
-        clientMode: require.resolve(
-          '../fixtures/custom-client/CustomSockJSClient'
-        ),
+        transportMode: {
+          server: 'sockjs',
+          client: require.resolve(
+            '../fixtures/custom-client/CustomSockJSClient'
+          ),
+        },
       },
     },
   ];
